requirements.txt
# 纳米酶文献提取系统 - 依赖包
# ============================

# Web/API 请求
aiohttp>=3.8.0
requests>=2.28.0

# 配置管理
pyyaml>=6.0

# 数据验证
pydantic>=2.0.0

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 开发工具（可选）
black>=23.0.0           # 代码格式化
flake8>=6.0.0           # 代码检查
mypy>=1.0.0             # 类型检查
